Created soft delete routes and controllers for daily_nutritions. Changed sleep field on daily_nutritions to integer.

+++ Backend/laravel-api/database/migrations/2025_10_17_213646_create_daily_nutrition_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             $table->integer('fat_consumed')->nullable();
             $table->float('water_intake')->nullable();
             $table->integer('steps')->nullable();
-            $table->float('sleep_hours')->nullable();
+            $table->integer('sleep_hours')->nullable();
             $table->integer('calories_burned')->nullable();
             $table->timestamps();
             $table->softDeletes();
